backend/requirements.txt
fastapi
uvicorn
python-multipart
sentence-transformers
faiss-cpu
numpy
pydantic-settings
python-dotenv
requests
llama-cpp-python
pdfplumber
beautifulsoup4
lxml
python-docx
duckduckgo-search
watchdog
pytest
psutil
